fix:新建useMenuHeight用来存储二级菜单高度，修复fuzzer缓存问题
