+++ app/renderer/src/main/src/store/pageInfo.ts
@@ -133,17 +133,11 @@ export interface ScanPortPageInfoProps {
     targets: string
 }
 
-interface MenuBodyHeightProps {
-    firstTabMenuBodyHeight: number
-}
+
 interface PageInfoStoreProps {
-    menuBodyHeight: MenuBodyHeightProps
     pages: Map<string, PageProps>
 
     selectGroupId: Map<string, string>
-
-    /**更新 menuBodyHeight*/
-    updateMenuBodyHeight: (v: MenuBodyHeightProps) => void
 
     /**设置 pages数据，例如：fuzzer缓存页面；未分组的关闭其他页面只保留当前页面*/
     setPagesData: (key: string, p: PageProps) => void
@@ -198,21 +192,10 @@ export const usePageInfo = createWithEqualityFn<PageInfoStoreProps>()(
     subscribeWithSelector(
         persist(
             (set, get) => ({
-                menuBodyHeight: {
-                    firstTabMenuBodyHeight: 0
-                },
+                
                 pages: new Map(),
                 selectGroupId: new Map(),
-                updateMenuBodyHeight: (value) => {
-                    const newVal = get().menuBodyHeight
-                    set({
-                        ...get(),
-                        menuBodyHeight: {
-                            ...newVal,
-                            ...value
-                        }
-                    })
-                },
+                
                 setPagesData: (key, values) => {
                     const newVal = new Map(get().pages).set(key, values)
                     set({
@@ -448,6 +431,7 @@ export const saveFuzzerCache = debounce(
                     color: ele.color
                 }
             })
+            console.log("save-cache", cache)
             setRemoteProjectValue(RemoteGV.FuzzerCache, JSON.stringify(cache)).catch((error) => {})
         } catch (error) {
             yakitNotify("error", "webFuzzer缓存数据失败:" + error)
